Combine reporting and stats tab into a cycle tab
