+++ src/AdminUI/src/app/components/subscriptions/manage-subscription/manage-subscription.component.html
@@ -5,11 +5,8 @@
   <mat-tab label="Subscription Configuration">
     <subscription-config-tab [showLoading]="loading"></subscription-config-tab>
   </mat-tab>
-  <mat-tab label="Statistics">
-    <subscription-stats-tab></subscription-stats-tab>
-  </mat-tab>
-  <mat-tab label="Reporting">
-    <subscription-report-tab></subscription-report-tab>
+  <mat-tab label="Cycles">
+    <app-subscription-cycles-tab></app-subscription-cycles-tab>
   </mat-tab>
   <mat-tab label="Tasks">
     <app-subscription-tasks-tab></app-subscription-tasks-tab>
