+++ src/AdminUI/src/app/components/subscriptions/manage-subscription/subscription-cycles-tab/subscription-cycles-tab.component.scss
@@ -4,5 +4,8 @@
 
 .cycle-select-wrapper {
   display: inline-block;
-  margin-left: 1em;
+}
+
+mat-card-title {
+  margin-bottom: 1rem;
 }
