--- src/AdminUI/src/app/components/subscriptions/manage-subscription/subscription-report-tab/subscription-report-tab.component.html
@@ -1,204 +1,0 @@
-<div *ngIf="subscription['cycles']">
-  <mat-card *ngIf="subscription['cycles']">
-    <mat-card-title> Reporting </mat-card-title>
-    Reporting Cycle:
-
-    <div class="cycle-select-wrapper" *ngIf="subscription">
-      <mat-form-field
-        class="cycle-select-dropdown"
-        *ngIf="subscription.cycles"
-        appearance="outline"
-      >
-        <mat-label>Cycle</mat-label>
-        <mat-select
-          [(value)]="selectedCycle"
-          (selectionChange)="cycleChange($event)"
-        >
-          <mat-option *ngFor="let c of subscription.cycles" [value]="c">
-            {{ c.start_date | UTCtoReadableTime }} -
-            {{ c.end_date | UTCtoReadableTime }} ({{ c.target_count }})
-          </mat-option>
-        </mat-select>
-      </mat-form-field>
-    </div>
-
-    <div class="flex flex-row">
-      Include Non-Human Interactions:
-      <mat-slide-toggle [(ngModel)]="includeNonhuman"></mat-slide-toggle>
-    </div>
-  </mat-card>
-  <mat-card>
-    <mat-label class="h6">Reports</mat-label>
-    <div class="mb-4">
-      <div
-        *ngIf="loading"
-        style="text-align: center; padding-top: 2rem; padding-bottom: 2rem"
-      >
-        <mat-spinner
-          class="spinner"
-          style="margin-left: auto; margin-right: auto"
-          diameter="50"
-        ></mat-spinner>
-      </div>
-      <div class="text-muted">View or send this subscription's reports</div>
-      <br /><br />
-      <button
-        mat-raised-button
-        color="primary"
-        class="mb-3"
-        (click)="downloadReport('status'); $event.preventDefault()"
-      >
-        Download Status Report
-      </button>
-      <button
-        mat-raised-button
-        color="primary"
-        class="mb-3"
-        (click)="downloadReport('cycle'); $event.preventDefault()"
-      >
-        Download Cycle Report
-      </button>
-      <button
-        mat-raised-button
-        color="gray"
-        class="mb-3"
-        [disabled]="true"
-        (click)="downloadReport('yearly'); $event.preventDefault()"
-      >
-        Download Yearly Report
-      </button>
-      <br /><br />
-      <button
-        mat-raised-button
-        color="accent"
-        class="mb-3"
-        (click)="sendReport('status'); $event.preventDefault()"
-      >
-        Send Status Report
-      </button>
-      <button
-        mat-raised-button
-        color="accent"
-        class="mb-3"
-        (click)="sendReport('cycle'); $event.preventDefault()"
-      >
-        Send Cycle Report
-      </button>
-      <button
-        mat-raised-button
-        color="gray"
-        class="mb-3"
-        [disabled]="true"
-        (click)="sendReport('yearly'); $event.preventDefault()"
-      >
-        Send Yearly Report</button
-      ><br />
-    </div>
-  </mat-card>
-  <mat-card>
-    <mat-label class="h6">Download Set Cycle Reports</mat-label>
-    <div>
-      <div
-        *ngIf="loading"
-        style="text-align: center; padding-top: 2rem; padding-bottom: 2rem"
-      >
-        <mat-spinner
-          class="spinner"
-          style="margin-left: auto; margin-right: auto"
-          diameter="50"
-        ></mat-spinner>
-      </div>
-      <div class="text-muted">
-        Download a report for a specified set of cycles
-      </div>
-      <button
-        mat-raised-button
-        color="primary"
-        class="mb-3"
-        (click)="openCycleSelectionDialog()"
-      >
-        Open Cycle Selection
-      </button>
-    </div>
-  </mat-card>
-  <mat-card>
-    <mat-label class="h6">Download Cycle Data</mat-label>
-    <div>
-      <div
-        *ngIf="loading"
-        style="text-align: center; padding-top: 2rem; padding-bottom: 2rem"
-      >
-        <mat-spinner
-          class="spinner"
-          style="margin-left: auto; margin-right: auto"
-          diameter="50"
-        ></mat-spinner>
-      </div>
-      <div class="text-muted">Download all data associated with this cycle</div>
-      <button
-        mat-raised-button
-        color="primary"
-        class="mb-3"
-        (click)="downloadCycleData(); $event.preventDefault()"
-      >
-        Download Data
-      </button>
-    </div>
-  </mat-card>
-
-  <!--
-    <div *ngIf="subscription">
-        <mat-form-field class="cycle-select-dropdown" *ngIf="subscription.cycles" appearance="outline">
-            <mat-label>Cycle</mat-label>
-            <mat-select  [(value)]="selectedCycle" (selectionChange)="cycleChange($event)">
-                <mat-option *ngFor="let c of subscription.cycles" [value]="c">
-                    {{ c.start_date | UTCtoReadableTime }} - {{ c.end_date | UTCtoReadableTime }}</mat-option>
-            </mat-select>
-        </mat-form-field>
-    </div> -->
-
-  <mat-card>
-    <div class="mb-3">
-      <mat-table [dataSource]="emailsSent" matSort>
-        <ng-container matColumnDef="report">
-          <mat-header-cell *matHeaderCellDef mat-sort-header
-            >Report</mat-header-cell
-          >
-          <mat-cell *matCellDef="let row">{{ row.message_type }}</mat-cell>
-        </ng-container>
-        <ng-container matColumnDef="sent">
-          <mat-header-cell *matHeaderCellDef mat-sort-header
-            >Sent</mat-header-cell
-          >
-          <mat-cell *matCellDef="let row">{{
-            row.sent | UTCtoReadableTime
-          }}</mat-cell>
-        </ng-container>
-        <ng-container matColumnDef="to">
-          <mat-header-cell *matHeaderCellDef mat-sort-header
-            >To</mat-header-cell
-          >
-          <mat-cell *matCellDef="let row">{{ row.email_to }}</mat-cell>
-        </ng-container>
-        <ng-container matColumnDef="from">
-          <mat-header-cell *matHeaderCellDef mat-sort-header
-            >From</mat-header-cell
-          >
-          <mat-cell *matCellDef="let row"> {{ row.email_from }}</mat-cell>
-        </ng-container>
-
-        <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row>
-        <mat-row
-          class="table-row"
-          *matRowDef="let row; columns: displayedColumns"
-        ></mat-row>
-      </mat-table>
-    </div>
-  </mat-card>
-</div>
-<div *ngIf="!subscription['cycles']" class="w-100">
-  <mat-card>
-    <mat-card-title> Reporting </mat-card-title>
-    Reporting will be available once a campaign has begun.
-  </mat-card>
-</div>
