--- src/AdminUI/src/app/components/subscriptions/manage-subscription/subscription-report-tab/subscription-report-tab.component.scss
@@ -1,30 +1,0 @@
-ng-select {
-  .ng-dropdown-panel {
-    width: auto !important;
-    max-width: 550px;
-    border: solid red 1px;
-  }
-}
-.cycle-select-dropdown {
-  min-width: 25em !important;
-}
-.cycle-select-wrapper {
-  display: inline-block;
-  margin-left: 1em;
-}
-
-.mat-column-to {
-  word-wrap: normal !important;
-  white-space: unset !important;
-  flex: 0 0 38% !important;
-  width: 48% !important;
-  overflow-wrap: break-word;
-  word-wrap: break-word;
-
-  word-break: break-word;
-
-  -ms-hyphens: auto;
-  -moz-hyphens: auto;
-  -webkit-hyphens: auto;
-  hyphens: auto;
-}
