--- src/AdminUI/src/app/components/subscriptions/manage-subscription/subscription-stats-tab/subscription-stats-tab.component.html
@@ -1,52 +1,0 @@
-<div *ngIf="subscription['cycles']" class="w-100">
-  <mat-card>
-    <mat-card-title> Statistics </mat-card-title>
-    Statistics Cycle:
-
-    <div class="cycle-select-wrapper" *ngIf="subscription">
-      <mat-form-field
-        class="cycle-select-dropdown"
-        *ngIf="subscription.cycles"
-        appearance="outline"
-      >
-        <mat-label>Cycle</mat-label>
-        <mat-select
-          [(value)]="selectedCycle"
-          (selectionChange)="cycleChange($event)"
-        >
-          <mat-option *ngFor="let c of subscription.cycles" [value]="c">
-            {{ c.start_date | UTCtoReadableTime }} -
-            {{ c.end_date | UTCtoReadableTime }} ({{
-              c.target_count
-            }})</mat-option
-          >
-        </mat-select>
-      </mat-form-field>
-    </div>
-    <button mat-raised-button color="primary" (click)="pageRefresh()">
-      Refresh
-      <mat-icon class="table-btn-icon-size">autorenew</mat-icon>
-    </button>
-
-    <div class="flex flex-row">
-      Include Non-Human Interactions:
-      <mat-slide-toggle (change)="showNonHuman($event)"></mat-slide-toggle>
-    </div>
-  </mat-card>
-  <app-sub-dashboard></app-sub-dashboard>
-
-  <mat-card *ngIf="display_timeline">
-    <div>
-      <h4>Subscription Timeline</h4>
-      <div style="margin: 0 2rem 2rem 2rem">
-        <app-svg-timeline [timelineItems]="timelineItems"></app-svg-timeline>
-      </div>
-    </div>
-  </mat-card>
-</div>
-<div *ngIf="!subscription['cycles']" class="w-100">
-  <mat-card>
-    <mat-card-title> Statistics </mat-card-title>
-    Statistics will be shown once a campaign has begun
-  </mat-card>
-</div>
