+++ src/AdminUI/src/styles.scss
@@ -374,8 +374,8 @@ $altTheme: mat-dark-theme($theme-primary, $theme-accent, $theme-warn);
 .mat-flat-button.mat-primary,
 .mat-fab.mat-primary,
 .mat-mini-fab.mat-primary {
-  color: #f0fff3 !important;
-  background-color: #29ab87 !important;
+  color: #f0fff3;
+  background-color: #29ab87;
 }
 
 .mat-icon-button.mat-primary {
